Backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
PyGithub==2.1.1
transformers>=4.41.0,<5.0.0
torch==2.1.0
langchain==0.0.335
python-magic==0.4.27
requests==2.31.0
numpy==1.24.3
pandas==2.1.1
scikit-learn==1.3.1
tqdm==4.66.1
gitpython==3.1.40
markdown==3.5
beautifulsoup4==4.12.2
sentence-transformers==4.1.0